minor changes based on coderabbit comments
