+++ includes/heredoc.h
@@ -6,7 +6,7 @@
 /*   By: ssoeno <ssoeno@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/06 13:31:55 by tamatsuu          #+#    #+#             */
-/*   Updated: 2025/01/19 00:55:33 by ssoeno           ###   ########.fr       */
+/*   Updated: 2025/01/19 15:29:24 by ssoeno           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,10 @@
 # include "./minishell.h"
 # include "./parser.h"
 
-int		input_heredoc_content(char *eof, t_context *ctx);
+// heredoc_read.c
+int		read_heredoc(char *eof, t_context *ctx);
+
+// heredoc.c
 void	heredoc_handler(t_node *node, t_context *ctx);
 void	call_heredoc(t_node *node, t_context *ctx);
 bool	is_heredoc_end(char **redirects);
