+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: ssoeno <ssoeno@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/25 21:31:55 by shokosoeno        #+#    #+#             */
-/*   Updated: 2025/01/19 01:01:00 by ssoeno           ###   ########.fr       */
+/*   Updated: 2025/01/19 15:54:20 by ssoeno           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,7 @@ typedef struct s_context {
 	t_map	*env;
 }	t_context;
 
-int		start_exec(char *line, t_context *ctx);
+void		start_exec(char *line, t_context *ctx);
 t_context	*init_ctx(void);
 void		clear_ctx(t_context *ctx);
