+++ srcs/heredoc/heredoc.c
@@ -6,7 +6,7 @@
 /*   By: ssoeno <ssoeno@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/05 02:14:42 by tamatsuu          #+#    #+#             */
-/*   Updated: 2025/01/19 15:18:54 by ssoeno           ###   ########.fr       */
+/*   Updated: 2025/01/19 15:27:55 by ssoeno           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,101 +15,6 @@
 #include <unistd.h>
 #include "../includes/heredoc.h"
 #include "../includes/signals.h"
-
-static int	terminate_heredoc(int *pipe_fds);
-int			read_heredoc_line(char *eof, int pipe_fd, t_context *ctx);
-
-// int	input_heredoc_content(char *eof, t_context *ctx)
-// {
-// 	char			*line;
-// 	int				pipe_fds[2];
-
-// 	line = NULL;
-// 	if (isatty(STDIN_FILENO))
-// 		rl_event_hook = heredoc_sigint_event_hook;
-// 	if (pipe(pipe_fds) == -1)
-// 		d_throw_error("input_heredoc_content", "pipe failed");
-// 	while (1)
-// 	{
-// 		free(line);
-// 		line = NULL;
-// 		line = readline("> ");
-// 		if (g_sig == SIGINT)
-// 		{
-// 			ctx->heredoc_interrupted = true;
-// 			return (terminate_heredoc(line, pipe_fds));
-// 		}
-// 		if (line == NULL || \
-// 		(ft_strlen(line) == ft_strlen(eof) && !ft_strcmp(eof, line)))
-// 			break ;
-// 		else
-// 		{
-// 			write(pipe_fds[1], line, ft_strlen(line));
-// 			write(pipe_fds[1], "\n", ft_strlen("\n"));
-// 		}
-// 	}
-// 	if (isatty(STDIN_FILENO))
-// 		rl_event_hook = sigint_event_hook;
-// 	free(line);
-// 	close(pipe_fds[1]);
-// 	return (pipe_fds[0]);
-// }
-/*
-by returning pipe_fds[0], we can redirect STDIN from this pipe
-so that the command reads these lines as input.
-*/
-
-int	input_heredoc_content(char *eof, t_context *ctx)
-{
-	int	pipe_fds[2];
-	int	ret;
-
-	if (isatty(STDIN_FILENO))
-		rl_event_hook = heredoc_sigint_event_hook;
-	if (pipe(pipe_fds) == -1)
-		d_throw_error("input_heredoc_content", "pipe failed");
-	ret = read_heredoc_line(eof, pipe_fds[1], ctx);
-	if (isatty(STDIN_FILENO))
-		rl_event_hook = sigint_event_hook;
-	close(pipe_fds[1]);
-	if (ret < 0)
-		return (terminate_heredoc(pipe_fds));
-	return (pipe_fds[0]);
-}
-
-int	read_heredoc_line(char *eof, int pipe_fd, t_context *ctx)
-{
-	char	*line;
-
-	line = NULL;
-	while (1)
-	{
-		free(line);
-		line = NULL;
-		line = readline("> ");
-		if (g_sig == SIGINT)
-		{
-			ctx->heredoc_interrupted = true;
-			free (line);
-			return (-2);
-		}
-		if (line == NULL || \
-		(ft_strlen(line) == ft_strlen(eof) && !ft_strcmp(eof, line)))
-			break ;
-		write(pipe_fd, line, ft_strlen(line));
-		write(pipe_fd, "\n", ft_strlen("\n"));
-	}
-	free(line);
-	return (EXIT_SUCCESS);
-}
-
-static int terminate_heredoc(int *pipe_fds)
-{
-	close(pipe_fds[1]);
-	if (isatty(STDIN_FILENO))
-		rl_event_hook = sigint_event_hook;
-	return (-2);
-}
 
 void	heredoc_handler(t_node *node, t_context *ctx)
 {
@@ -137,7 +42,7 @@ void	call_heredoc(t_node *node, t_context *ctx)
 	{
 		if (!ft_strcmp(node->redirects[i], "<<") && node->redirects[i + 1])
 		{
-			temp_fd_arry[j++] = input_heredoc_content(node->redirects[i + 1], ctx);
+			temp_fd_arry[j++] = read_heredoc(node->redirects[i + 1], ctx);
 			i++;
 		}
 		i++;
